refactor: moved actions pkg to api
